+++ docs/changelog.md
@@ -122,3 +122,11 @@
 - Added code to SnakeView that calls the SnakeMusic class.
 - Added a proper end screen which show the final score, and has a countdown to when the game will restart.
 - Updated how the game ending is handled, moved to moveSnake method.
+
+### 10/12/23
+- Added mechanism for restarting the game in Controller. Method restartGame.
+- Added method to view restartGame that resets the UI elements.
+- Added two buttons to the end screen to restart the game and to go back to the main menu. (Not yet implemented.)
+- Added CSS to the buttons to style them.
+- Commented CSS.
+- Implemented Restart Button.